Got the system to clear regression tests again, now it is actually doing some calculations. The tolerance is rather loose at 0.2%, and I would like to find better reference output.
git-svn-id: https://svn.code.sf.net/p/moose/code/moose/branches/Msg@2244 d1858a74-f715-0410-99c9-c66fe9cce3d6

+++ ksolve/Stoich.cpp
@@ -51,7 +51,7 @@ const Cinfo* Stoich::initCinfo()
 			&Stoich::getOneWay
 		);
 
-		static ReadOnlyValueFinfo< Stoich, double > nVarMols(
+		static ReadOnlyValueFinfo< Stoich, unsigned int > nVarMols(
 			"nVarMols",
 			"Number of variable molecules in the reac system",
 			&Stoich::getNumVarMols
@@ -185,7 +185,7 @@ string Stoich::getPath( const Eref& e, const Qinfo* q ) const
 	return path_;
 }
 
-double Stoich::getNumVarMols() const
+unsigned int Stoich::getNumVarMols() const
 {
 	return numVarMols_;
 }
